added authorization filters

+++ app/controllers/secrets_controller.rb
@@ -15,6 +15,7 @@ class SecretsController < ApplicationController
 
   def destroy
     secret = Secret.find(params[:id])
+    redirect_to "/users/#{current_user.id}" unless current_user.id == secret.user_id
     secret.destroy if secret.user === current_user
     redirect_to "/users/#{current_user.id}"
   end
